[AutoDiff] Split auto_diff.cpp into stage-grouped files under quadrants/transforms/auto_diff/
